Brillouin Zone

MonkhorstPack

WannierExcitonModel.MonkhorstPack — Type

MonkhorstPack <: AbstractBrillouinZone

Perform BZ sampling employing a Monkhorst-Pack grid.

julia> kgrid = MonkhorstPack([5, 5, 1]; kshift = [1//2, 0, 0])
MonkhorstPack([5, 5, 1], [0.5, 0.0, 0.0])

julia> length(kgrid)
25

RedKgrid

WannierExcitonModel.RedKgrid — Type

RedKgrid <: AbstractBrillouinZone

Actually it's a simple expansion of MonkhorstPack. It supports bracket-based access.

julia> kgrid = MonkhorstPack([5, 5, 1]; kshift = [1//2, 0, 0])
MonkhorstPack([5, 5, 1], [0.5, 0.0, 0.0])

julia> kgrid = RedKgrid(kgrid)
Reducibal kgrids with 25 reducible k-points.

julia> length(kgrid)
25

julia> kgrid[2]
3-element ReducedCoordinates{Rational{Int64}} with indices SOneTo(3):
 -3//10
 -2//5
   0

Kline

WannierExcitonModel.Kline — Type

Kline{L <: Real} <: AbstractBrillouinZone

Kline includes some useful fileds in order to calculate band structure.

line::Vector{Float64}: used as the horizontal coordinate of the band structure;

name::Vector{String}: the name of the high symmetry point;

index::Vector{Int}: the name of the high symmetry point.

It also supports bracket-based access.

julia> kline[3]

WannierExcitonModel.Kline — Method

Kline(nk::Integer, TB::AbstractTightBindModel/cell::Cell) -> Kline
Kline(hspoint::AbstractVector{<:Pair}, nk::Integer, lattice::Lattice/cell::Cell/TB::AbstractTightBindModel) -> Kline

For the first method, Kline will try to build a high symmetry path in FBZ. But we have only implemented a few 2D scenarios at present, and we also support the automatic generation of Kline in 1D and 0D.

For other case, you need to provide high symmtry points by hspoint. For example:

hspoint = [
	"L" => [0.5, 0.5, 0.5],
	"G" => [0, 0, 0],
	"X" => [0.5, 0, 0.5],
	"W" => [0.5, 0.25, 0.75],
	"L" => [0.5, 0.5, 0.5],
]

kgridmap

WannierExcitonModel.kgridmap — Function

kgridmap(aimkdirects, kdirects, mode = +) -> Matrix{Int}

Try to get a map matrix between aimkdirects and kdirects.

julia> M = kgridmap(aimkdirects, kdirects, +);

julia> aimkdirects[M[i, j]] == mode(kdirects[i], kdirects[j])
true

We also provide some shortcuts:

kgridmap(kgrid::MonkhorstPack, mode = +)
kgridmap(redkgrid::RedKgrid, mode = +)

fold2FBZ

WannierExcitonModel.fold2FBZ — Function

fold2FBZ(Rlattice::ReciprocalLattice, kpoints::AbstractVector{<:AbstractVector}) -> AbstractVector{<:AbstractVector}
fold2FBZ(Rlattice::ReciprocalLattice, kpoint::AbstractVector{<:Real}) -> AbstractVector{<:Real}

Return the input's equivalent kpoints in FBZ.