Coulomb

We provide some types to calculate different coulomb potentials.

Real space

WannierExcitonModel.RealInverseR — Type

RealInverseR(; ϵ::Real = 1) -> RealInverseR

Calculate Coulomb potential generated by a point charge in real space:

\[V(r) = \frac{e²}{4πϵ₀ϵ} \frac{1}{|r|}\]

unit: r(Å), V(eV), ϵ(dimensionless).

Example

julia> V = RealInverseR(; ϵ = 1.5)
julia> r = [1.0, 0.0, 0.0] # or r = 1.0
julia> V(r)

Note V(0) = 0 here.

WannierExcitonModel.RealGauss — Type

RealGauss(; ϵ::Real = 1, α::Real = 1) -> RealGauss

Calculate Coulomb potential generated by a charge of Gaussian distribution in real space:

\[V(r) = \frac{e²}{4πϵ₀ϵ} \frac{erf(α|r|)}{|r|}\]

unit: r(Å), V(eV), α(1/Å), ϵ(dimensionless).

Example

julia> V = RealGauss(; ϵ = 1.5, α = 2)
julia> r = [1.0, 0.0, 0.0] # or r = 1.0
julia> V(r)

Reciprocal space

WannierExcitonModel.ReciprocalInverseR — Type

ReciprocalInverseR(; ϵ::Real = 1, Ω::Real = 1) -> ReciprocalInverseR

Calculate Coulomb potential generated by a point charge in reciprocal space:

\[V(k) = \frac{e²}{Ωϵ₀ϵ} \frac{1}{k²}\]

unit: k(1/Å), V(eV), Ω(Å³), ϵ(dimensionless).

Example

julia> V = ReciprocalInverseR(; ϵ = 1.5, Ω = 10.0)
julia> k = [1.0, 0.0, 0.0] # or k = 1.0
julia> V(k)

We provide a method to calculate the head value:

\[V(head) = \frac{NΩ}{(2π)³} \int_{0}^{q_{sz}} V(k) 4πk² dk , \frac{4π}{3}q_{sz}^3 = \frac{(2π)³}{NΩ}\]

julia> V(:head, N)

WannierExcitonModel.ReciprocalGauss3D — Type

ReciprocalGauss3D(; ϵ::Real = 1, α::Real = 1, Ω::Real = 1) -> ReciprocalGauss3D

Under 3D cases, calculate Coulomb potential generated by a charge of Gaussian distribution in reciprocal space:

\[V(k) = \frac{e²}{Ωϵ₀ϵ} \frac{1}{k²} e^{-\frac{k²}{4α²}}\]

unit: k(1/Å), V(eV), Ω(Å³), α(1/Å), ϵ(dimensionless).

Example

julia> V = ReciprocalGauss3D(; ϵ = 1.5, α = 1.0, Ω = 10.0)
julia> k = [1.0, 0.0, 0.0] # or k = 1.0
julia> V(k)

We provide a method to calculate the head value:

\[V(head) = \frac{NΩ}{(2π)³} \int_{0}^{q_{sz}} V(k) 4πk² dk , \frac{4π}{3}q_{sz}^3 = \frac{(2π)³}{NΩ}\]

julia> V(:head, N)

WannierExcitonModel.ReciprocalGauss2D — Type

ReciprocalGauss2D(; ϵ::Real = 1, α::Real = 1, S::Real = 1) -> ReciprocalGauss2D

Under 2D cases, calculate Coulomb potential generated by a charge of Gaussian distribution in reciprocal space:

\[V(k, z) = \frac{e²}{4Sϵ₀ϵ} \frac{1}{k} \left[ e^{kz} erfc(\frac{k}{2α} + αz) + e^{-kz} erfc(\frac{k}{2α} - αz) \right]\]

unit: k(1/Å), z(Å), V(eV), S(Å²), α(1/Å), ϵ(dimensionless).

Example

julia> V = ReciprocalGauss2D(; ϵ = 1.5, α = 1.0, S = 10.0)
julia> k = [1.0, 0.0, 0.0] # or k = 1.0
julia> z = 0.5
julia> V(k, z)

We provide a method to calculate the head value:

\[V(head, z) = \frac{NS}{(2π)²} \int_{0}^{q_{sz}} V(k) 2πk dk , πq_{sz}^2 = \frac{(2π)²}{NS}\]

julia> V(:head, N, z)