Kernal

WannierExcitonModel.AbstractKernalInterAction — Type

AbstractKernalInterAction

This abstract type is used to calculate kernal term in excitonic BSE model. You can design a concrete type of AbstractKernalInterAction according to your own needs.

For YourKernal, it should have specific fields:

nk::Int: the number of kpoints;

kgrid::RedKgrid: the kgrid used to calculate excitonic BSE model;

kgrid_Γ::RedKgrid: with the same size of kgrid but without shift;

kgrid_addmap::Matrix{Int}: satisfy kgrid[i] + kgrid[j] = kgrid_Γ[kgrid_addmap[i, j]]

kgrid_minusmap::Matrix{Int}: satisfy kgrid[i] - kgrid[j] = kgrid_Γ[kgrid_minusmap[i, j]]

YourKernal should support these methods:

(K::YourKernal)(::Val{:initialize})

It's used to calculate some interaction matrix at first, for example W(k′-k); it will be run when create a BSE model.

(K::YourKernal)(q::ReducedCoordinates)

It's used to calculate interaction matrix related to q, for example V(q) or J(k+k′+q); it will be run when calculate BSE Hamiltonian.

(K::YourKernal)(k′, k, ψ₁, ψ₂, ψ₃, ψ₄)

It's used to calculate kernal matrix elements $K_{v'c'k',vck}$,

k′ and k: it's the index of kpoints in kgrid list;

ψ: is a numerical vector; from left to right, in sequence: $ψ_{c′}^{k′+q}$, $ψ_{v′}^{k′}$, $ψ_{v}^{k}$, $ψ_{c}^{k+q}$.

Note

You can change inputs, but use local operation carefully. Do not forget the minus of Kᵈ.

For the case only refer to q in a specified qgrid, i.e. isqgrid is true, this also means k+q is in the kgrid for any q needed. In this case, YourKernal needs to support two more methods:

(K::YourKernal)(::Val{:initialize_qgrid})

It's is a supplement to (K::YourKernal)(::Val{:initialize}); only run after (K::YourKernal)(::Val{:initialize}) when create BSE model.

(K::YourKernal)(kq_kindex::Vector{Int}, kΓq_kΓindex::Vector{Int}, q::ReducedCoordinates)

When isqgrid is true, BSEqgrid will not use (K::Kernal)(q::ReducedCoordinates) except when q equal to [0, 0, 0], instead of using this function.

kq_kindex::Vector{Int}: kgrid[i] + q = kgrid[kq_kindex[i]];

kΓq_kΓindex::Vector{Int}: kgrid_Γ[i] + q = kgrid_Γ[kq_kindex[i]];

UJ approximation

WannierExcitonModel.Kernal_UJ — Method

Kernal_UJ(kgrid::Union{MonkhorstPack, RedKgrid}, 
	Kᵈ_U::UwithLR{UT_d, CT_d}, Kᵈ_J¹::AbstractReciprocalHoppings, Kᵈ_J²::AbstractReciprocalHoppings,
	Kˣ_U::UwithLR{UT_x, CT_x}, Kˣ_J¹::AbstractReciprocalHoppings, Kˣ_J²::AbstractReciprocalHoppings;
	upindex::Union{<:AbstractVector{<:Integer}, Nothing} = nothing,
	dnindex::Union{<:AbstractVector{<:Integer}, Nothing} = nothing
) -> AbstractKernalInterAction

Creat a object which can be provided to canstruct a BSE model.

kgrid: should be the same as excitonic basis;

K_U: direct Coulomb interaction with long-range correction between electronic wannier basis;

K_J: exchange Coulomb interaction between electronic wannier basis;

up(dn)_index: the wannier basis index when contains spin.

Kᵈ needs screened coulomb interaction between wannier basis, while Kˣ needs naked coulomb interaction. The long-range behavior of direct Coulomb interaction between wannier basis is proportional to \frac{1}{r}, we can't dismiss it. Check UwithLR to see how to contain the long-range correction. See more details about J¹ and J² in Theory.

For spinless case, you shouldn't provide upindex and dnindex. But for spinful case, you must provide at least one between upindex and dnindex. For spinful case, note that the interaction should not contain spin(see more details in Theory.)

Kernal_UJ(kgrid::Union{MonkhorstPack, RedKgrid}, 
	Kᵈ_U::UwithLR, Kᵈ_J¹::Union{AbstractString, HR}, Kᵈ_J²::Union{AbstractString, HR},
	Kˣ_U::UwithLR, Kˣ_J¹::Union{AbstractString, HR}, Kˣ_J²::Union{AbstractString, HR};
	upindex::Union{<:AbstractVector{<:Integer}, Nothing} = nothing,
	dnindex::Union{<:AbstractVector{<:Integer}, Nothing} = nothing
) -> AbstractKernalInterAction

This is a shortcut.

Note

For spinful case, we require that the electronic wannier basis satisfy time-reversal symmetry, so that we can define which term is U and which term is J.

U approximation

WannierExcitonModel.Kernal_U — Method

Kernal_U(kgrid::Union{MonkhorstPack, RedKgrid}, 
	Kᵈ_U::UwithLR, Kˣ_U::UwithLR;
	upindex::Union{<:AbstractVector{<:Integer}, Nothing} = nothing,
	dnindex::Union{<:AbstractVector{<:Integer}, Nothing} = nothing
) -> AbstractKernalInterAction

Creat a object which can be provided to canstruct a BSE model.

kgrid: should be the same as excitonic basis;

K_U: direct Coulomb interaction with long-range correction between electronic wannier basis;

up(dn)_index: the wannier basis index when contains spin.

Its usage is similar to Kernal_UJ.

Note

For spinful case, we require that the electronic wannier basis satisfy time-reversal symmetry, so that we can define which term is U.