Mirror Correction

WannierExcitonModel.MirrorCorrection — Method

MirrorCorrection(U::HR, TB::TightBindModel, rcut::Real; kwargs...) -> HR

Try to apply a mirror correction to U. This mirror correction is achieved by using Gauss potential.

rcut::Real: should be less than a real radius value, out of which the U approximate to 1/r, default is judged by U and TB;

kwargs:

kgrid::MonkhorstPack: should be the kgrid used to calculate U, default is judged by U;

ϵ::Real: dielectric constant;

αrcut::Real: α × rcut, default is 4.5;

δ:Real: the cutoff value of reciprocal potential.