Reciprocal Hopping

This part is used to process the hopping terms of a tight binding model.

HR

WannierExcitonModel.HR — Type

HR{T <: Number}

The data in wannier90_hr.dat. Fields:

orbindex::Vector{Int}: the index of orbtal, usually equal to 1:norb;

path::Matrix{Int}: its dimension is N × 5, each row is [Rx, Ry, Rz, i, j];

value::Vector{T}: its length is N.

construct

WannierExcitonModel.HR — Method

HR(path::AbstractMatrix{<:Integer}, value::AbstractVector{T}; orbindex = sort(unique(path[:, 4:5])), μ::Real = 0, hrsort = 'N') -> HR{T}

Create a HR object from path and value.

path is a N × 5 matrix;

value is a vector of length N;

orbindex is the orbital index list, default is as follows;

μ is the fermi energy shift, default is 0.

hrsort is whether to sort the hr by path, default is 'N', can be 'Y' or 'N'.

Hopping

WannierExcitonModel.Hopping — Type

Hopping{T <: Number}

A Hopping represents a hopping term in a tight-binding model, defined by an initial orbital index i in unitcell [0, 0, 0], a final orbital index j in unitcell R (in reduced coordinates), and a hopping value t of type T.

julia> hop = Hopping([1, 2, 3, 4, 5], 0.5)
t₄₅[[1, 2, 3]] = 0.5

julia> hop = Hopping(4, 5, [1, 2, 3], 0.5)
t₄₅[[1, 2, 3]] = 0.5

julia> path(hop)
3-element ReducedCoordinates{Int64} with indices SOneTo(3):
 1
 2
 3

julia> value(hop)
0.5

julia> k = ReducedCoordinates([0.1, 0.2, 0.3]);

julia> hopphase(hop, k) == exp(1im * 2π * (k ⋅path(hop)))
true

AbstractReciprocalHoppings

WannierExcitonModel.AbstractReciprocalHoppings — Type

AbstractReciprocalHoppings{T <: Number}

Reorganized the storage form of the hopping terms to facilitate calculation.

BaseReciprocalHoppings{T} <: AbstractReciprocalHoppings{T}

A concrete type of AbstractReciprocalHoppings. But don't use BaseReciprocalHoppings directly.

ReciprocalHoppings

WannierExcitonModel.ReciprocalHoppings — Type

ReciprocalHoppings{T, U <: BaseReciprocalHoppings{T}} <: AbstractReciprocalHoppings{T}

julia> hr isa HR
true

julia> rh = ReciprocalHoppings(hr);

julia> k isa ReducedCoordinates
true

julia> rh(k);

HermitianReciprocalHoppings

WannierExcitonModel.HermitianReciprocalHoppings — Type

HermitianReciprocalHoppings{T <: Number, U <: AbstractReciprocalHoppings{T}} <: AbstractReciprocalHoppings{T}

julia> hr isa HR
true

julia> hrh = HermitianReciprocalHoppings(hr);

julia> k isa ReducedCoordinates
true

julia> hrh(k) isa Hermitian
true