Structure

We provide a group of types to help you build a crystalline structure. We draw inspiration from CrystallographyCore.jl to design some types. For most cases, these are only intermediate objects, and we use them to construct a Model type.

Lattice

It can be used just like a matrix.

WannierExcitonModel.Lattice — Method

Lattice(data::AbstractMatrix) -> Lattice

Construct a Lattice from a matrix.

Note

The basis vectors of the matrix are stored as columns.

Examples

julia> Lattice([
		   1.2 4.5 7.8
		   2.3 5.6 8.9
		   3.4 6.7 9.1
	   ])
Lattice{Float64}
 1.2  4.5  7.8
 2.3  5.6  8.9
 3.4  6.7  9.1

WannierExcitonModel.Lattice — Method

Lattice(𝐚::AbstractVector, 𝐛::AbstractVector, 𝐜::AbstractVector) -> Lattice

Construct a Lattice from three basis vectors.

Examples

julia> 𝐚, 𝐛, 𝐜 = [1.2, 2.3, 3.4], [4.5, 5.6, 6.7], [7.8, 8.9, 9.10];

julia> Lattice(𝐚, 𝐛, 𝐜)
Lattice{Float64}
 1.2  4.5  7.8
 2.3  5.6  8.9
 3.4  6.7  9.1

Coordinates

They can be created and used just like a vector.

WannierExcitonModel.ReducedCoordinates — Type

Reduced Coordinates.

WannierExcitonModel.CrystalCoordinates — Type

CrystalCoordinates = ReducedCoordinates

WannierExcitonModel.CartesianCoordinates — Type

Cartesian Coordinates.

Example

julia> using WannierExcitonModel
julia> lattice = Lattice([
       		   1.2 4.5 7.8
       		   2.3 5.6 8.9
       		   3.4 6.7 9.1
       	   ])Lattice{Float64}
 1.2  4.5  7.8
 2.3  5.6  8.9
 3.4  6.7  9.1
julia> a = ReducedCoordinates(1, 2, 3)3-element ReducedCoordinates{Int64} with indices SOneTo(3):
 1
 2
 3
julia> b = lattice * a3-element CartesianCoordinates{Float64} with indices SOneTo(3):
 33.6
 40.2
 44.1
julia> lattice \ b3-element ReducedCoordinates{Float64} with indices SOneTo(3):
 1.0000000000000135
 2.000000000000027
 2.999999999999997

Cell

WannierExcitonModel.Cell — Type

Cell{P <: Union{CartesianCoordinates{<:Real}, ReducedCoordinates{<:Real}}}

Contains all the information of a crystalline unitcell. Fields:

lattice::Lattice{Float64}: lattice basis matrix;

location::Vector{P}: all atoms' location;

name::Vector{String}: all atoms' location;

index::Vector{Int}: all atoms' index;

period::Vec3{Bool}: the periodicty of unitcell.

construct

WannierExcitonModel.Cell — Method

Cell(lattice, location, location_type = "Cartesian"; name = String[], index = Int[], period = Bool[1, 1, 1])

Create a new cell.

lattice: a Lattice type or a matrix, note the basis vectors of the matrix are stored as columns.

location: a vector contains all atom's location, its length equal to $N$;

location_type: "Reduced" or "Cartesian";

name: a list of $N$ values, where the same kind of name need to be a String.

index: a list of $N$ values, where the same kind of index need to be a Integer.

period: [1(0), 1(0), 1(0)].

Make sure the basis at unperiodic direction is perpendicular to the other basis.

ORBITAL

WannierExcitonModel.ORBITAL — Type

ORBITAL{OT <: Real, AT <: Real}

The data in wannier90_centres.xyz. Fields:

location::Vector{CartesianCoordinates{OT}}: location of orbital centres;

name::Vector{String}: name of orbitals;

index::Vector{Int}: index of orbitals;

atom_location::Vector{CartesianCoordinates{AT}}: location of atoms;

atom_name::Vector{String}: name of atoms;

belonging::Vector{Int}: the index of atom that orbital belongs to.

construct

WannierExcitonModel.ORBITAL — Method

ORBITAL(location::AbstractVector; name = String[], index = Int[], 
	atom_location = CartesianCoordinates{Float64}[], atom_name = String[], belonging = Int[])

Create an ORBITAL object.

location: a vector contains all the locations of orbitals, its length equals $N$;

name and index: its element is the name or index of each orbital;

atom_location and atom_name: its element is the location or name of each orbital;

belonging: the i-th orbital belongs to the belonging[i]-th atom.