Tight Binding Model

This package provides a group of functions to process tight-binding model. The core object is TightBindModel,

WannierExcitonModel.TightBindModel — Type

TightBindModel{T <: Number, U <: AbstractReciprocalHoppings{T}} <: AbstractTightBindModel

A tight binding model object containing fields:

lattice::Lattice{Float64}: Lattice of the system.

atom_name::Vector{String}: Names of atoms of the unit cell.

atom_location::Vector{ReducedCoordinates{Float64}}: Locations of atoms in reduced coordinates.

orb_name::Vector{String}: Names of orbitals of the unit cell.

orb_location::Vector{ReducedCoordinates{Float64}}: Locations of orbitals in reduced coordinates.

H::HermitianReciprocalHoppings{T, U}: Hermitian Hamiltonian of the tight binding model.

period::Vec3{Bool}: Periodicity of the system in each lattice basis direction.

Note that the auxiliary basis of lattice in unperiodic directions should be vertical to the periodic directions.

numatom(TB::TightBindModel) -> Int
numorb(TB::TightBindModel) -> Int

Used to obtain the number of atoms or orbitals.

Construct

You can construct a TightBindModel by

WannierExcitonModel.TightBindModel — Method

TightBindModel(;
	hops::Union{AbstractString, HR, AbstractReciprocalHoppings},
	cell::Union{AbstractString, Cell},
	orbital::Union{AbstractString, ORBITAL},
	period = Bool[1, 1, 1],
)

Create an TightBindModel object. The AbstractString is the path to aim file, and check docstrings of HR, AbstractReciprocalHoppings, Cell and ORBITAL.

We emphasize that there are three parts of TightBindModel are the important. They are:

common file format

lattice basis vector	POSCAR.vasp

orbital center location	wannier90_centres.xyz

hopping terms	wannier90_hr.dat

The simplest method to construct a TightBindModel is to run:

julia> TightBindModel(;	hops = "wannier90_hr.dat", cell = "POSCAR.vasp", orbital = "wannier90_centres.xyz", period)

Remember to set period.

Hamiltonian

If you already have a instance TB of type TightBindModel, then you can run:

julia> k = ReducedCoordinates(1, 2, 3)
julia> TB(k)

to get Hamiltonian of the tight-binding model. Or you may want to store Hamiltonian matrix to a existing variable:

julia> k = ReducedCoordinates(1, 2, 3)
julia> H = Matrix{Float64}(undef, numorb(TB), numorb(TB))
julia> TB(H, k)

When you need the Hamiltonian under atom guage, whose eigen vectors represent $\left| u_{n\mathbf{k}} \right\rangle$(see more details in Theory), you can run:

julia> TB(k, TB.orb_location)
julia> TB(H, k, TB.orb_location)

Under atom guage, you can also get the partial derivative of Hamiltonian from

julia> TB(:partial, k)
julia> TB(:partial, H, k)
julia> size(H) == (numorb(TB), numorb(TB), 3)
true

BAND

The function BAND is the core of this package.

WannierExcitonModel.BAND — Method

BAND(kpoints, TB::AbstractTightBindModel; vector::Bool = false, wfctype::Symbol = :Bloch)

This method can calculate the band structure of a tight-binding model.

kpoints can be a AbstractBrillouinZone object, a vector of kpoints or a single kpoint, where kpoint should be a vector of 3 real numbers.

vector can be set as true, which decides whether to output wavefunction.

wfctype can be set as `:Periodic, which decides the type of wavefunction.

WannierExcitonModel.ExtractU — Method

ExtractU(kpoints, band, TB::AbstractTightBindModel) -> band_u
ExtractU(kpoints, band, orblocat::AbstractVector{<:ReducedCoordinates}) -> band_u

This function can extract the periodic part of the wavefunction from the bloch wavefunction.

kpoints can be a AbstractBrillouinZone object, a vector of kpoints or a single kpoint, where kpoint should be a vector of 3 real numbers.

band can be a vector of Eigen object or a single Eigen object.

band_u is a vector of Eigen object or a single Eigen object depending on the input band.

Expectation of operator

You can calculate the expectation value of an operator for each eigenstate, with the matrix representation of this operator.

WannierExcitonModel.ExpectationValue — Method

ExpectationValue(operator::AbstractMatrix{<:Number}, band::AbstractVector{<:Eigen}) -> Vector{Vector{Float64}}
ExpectationValue(operator::AbstractMatrix{<:Number}, band::Eigen) -> Vector{Float64}

Calculate the expectation value of an operator for each eigenstate.

spin operator

For the simplest case where the tight-binding model contains spin, we can obtain the $\hat{S}_{z}$ matrix by:

julia> numorb(TB)
6
julia> TB(:spinmat, [1, 2, 3]) # equal to TB(:spinmat, [1, 2, 3], [4, 5, 6])

The unit is $\frac{\hbar}{2}$.