UwithLR

WannierExcitonModel.AbstractLRCorrection — Type

abstract type AbstractLRCorrection end

Long-range correction for coulomb interaction, can give the potential matrix in reciprocal space. It supports

(v::AbstractLRCorrection)(k::AbstractVector)
(v::AbstractLRCorrection)(k::AbstractVector, nk::Integer)
(v::AbstractLRCorrection)(A::AbstractMatrix, k::AbstractVector)
(v::AbstractLRCorrection)(A::AbstractMatrix, k::AbstractVector, nk::Integer)
(v::AbstractLRCorrection)(::Val{+}, A, k::AbstractVector)
(v::AbstractLRCorrection)(::Val{+}, A, k::AbstractVector, nk::Integer)

The methods involving nk can give head term when k is zero in all periodic directions.

WannierExcitonModel.UwithLR — Type

UwithLR{<: AbstractReciprocalHoppings, <: AbstractLRCorrection} <: AbstractInterAction

Direct coulomb term between wannier basis with long-range correction. Fields:

norb::Int: the number of orbitals;

SR::AbstractReciprocalHoppings: short-range part of direct term;

LR::AbstractLRCorrection: long-range part of direct term.

We have achieved long-range correction by using Gauss potential. If you want to use other potential, you can define a new subtype of AbstractLRCorrection. Make sure your subtype has methods:

(v::yoursubtype)(::Val{+}, A, k::AbstractVector)
(v::yoursubtype)(::Val{+}, A, k::AbstractVector, nk::Integer)

For an instance u of UwithLR, you can run:

julia> u(k)
julia> u(k, nk)

These two methods will create a new matrix. And u(k, nk) will judge whether head term or normal value by k.

julia> u(A, k)
julia> u(A, k, nk)

These two methods will change A instead of creating a new matrix.

construct

WannierExcitonModel.UwithLR — Method

UwithLR(U, TB::TightBindModel, rcut::Real; kwargs...) -> UwithLR

A shortcut to create a UwithLR. This long-range correction is achieved by using Gauss potential.

U: the direct Coulomb potential between the wannier bases within a limited distance, can be a path::String, hr::HR or rh::AbstractReciprocalHoppings.

kwargs:

ϵ::Real: dielectric constant;

αrcut::Real: α × rcut, default is 4.5;

δ:Real: the cutoff value of reciprocal potential.